Implemented bearer token reuse and expiry
Removed the generate_token method of Telstra::SMS
Added a private authorize method to fetch the OAuth token and expiry
Added a private update_token method to fetch the token and
  set the token and token_expires instance variables
Added a token_expired? method to test whether the token is valid
Removed the token accessor and added a token method to
  update the token if required and return the bearer token

Updated Telstra::SMS spec with token and token_exprired? methods